only owner of contact can delete contact

+++ app/Http/Controllers/ContactController.php
@@ -97,6 +97,9 @@ class ContactController extends Controller
      */
     public function destroy(Contact $contact)
     {
+        if(\request()->user()->isNot($contact->user)){
+            return \response([],403); 
+        }
         $contact->delete();
     }
 }
